+++ frontend/src/metabase/actions/components/ActionForm/ActionForm.styled.tsx
@@ -1,12 +1,7 @@
 import styled from "@emotion/styled";
-import { space } from "metabase/styled-components/theme";
 
 export const ActionFormButtonContainer = styled.div`
   display: flex;
   justify-content: flex-end;
   gap: 0.5rem;
-`;
-
-export const FormFieldEditorDragContainer = styled.div`
-  margin-bottom: ${space(1)};
 `;
