+++ frontend/src/metabase/actions/containers/ActionCreator/FormCreator/FormCreator.tsx
@@ -1,21 +1,47 @@
-import React, { useState, useEffect, useMemo } from "react";
+import React, { useEffect, useCallback, useMemo, useState } from "react";
 import { jt, t } from "ttag";
 import _ from "underscore";
+import { DragDropContext, Droppable, Draggable } from "react-beautiful-dnd";
+
+import type {
+  DraggableProvided,
+  DroppableProvided,
+  DropResult,
+} from "react-beautiful-dnd";
 
 import ExternalLink from "metabase/core/components/ExternalLink";
+import Form from "metabase/core/components/Form";
+import FormProvider from "metabase/core/components/FormProvider";
 
 import MetabaseSettings from "metabase/lib/settings";
 
-import { ActionForm } from "metabase/actions/components/ActionForm";
 import SidebarContent from "metabase/query_builder/components/SidebarContent";
 
-import type { ActionFormSettings, Parameter } from "metabase-types/api";
+import type {
+  ActionFormSettings,
+  FieldSettings,
+  Parameter,
+} from "metabase-types/api";
 
-import { getDefaultFormSettings, sortActionParams } from "../../../utils";
+import {
+  getForm,
+  getDefaultFormSettings,
+  sortActionParams,
+} from "../../../utils";
 import { syncFieldsWithParameters } from "../utils";
+import { reorderFields } from "./utils";
 
 import { EmptyFormPlaceholder } from "./EmptyFormPlaceholder";
-import { FormContainer, InfoText } from "./FormCreator.styled";
+import FormFieldEditor from "./FormFieldEditor";
+import {
+  FormContainer,
+  FormFieldEditorDragContainer,
+  InfoText,
+} from "./FormCreator.styled";
+
+// FormEditor's can't be submitted as it serves as a form preview
+const BLANK_INITIAL_VALUES = {};
+const ON_SUBMIT_NOOP = _.noop;
 
 interface FormCreatorProps {
   parameters: Parameter[];
@@ -45,9 +71,53 @@ function FormCreator({
     }
   }, [parameters, formSettings]);
 
+  const form = useMemo(
+    () => getForm(parameters, formSettings?.fields),
+    [parameters, formSettings?.fields],
+  );
+
   const sortedParams = useMemo(
     () => parameters.sort(sortActionParams(formSettings)),
     [parameters, formSettings],
+  );
+
+  const handleDragEnd = useCallback(
+    ({ source, destination }: DropResult) => {
+      if (!formSettings.fields) {
+        return;
+      }
+
+      const oldOrder = source.index;
+      const newOrder = destination?.index ?? source.index;
+
+      const reorderedFields = reorderFields(
+        formSettings.fields,
+        oldOrder,
+        newOrder,
+      );
+      setFormSettings({
+        ...formSettings,
+        fields: reorderedFields,
+      });
+    },
+    [formSettings],
+  );
+
+  const handleChangeFieldSettings = useCallback(
+    (newFieldSettings: FieldSettings) => {
+      if (!newFieldSettings?.id) {
+        return;
+      }
+
+      setFormSettings({
+        ...formSettings,
+        fields: {
+          ...formSettings.fields,
+          [newFieldSettings.id]: newFieldSettings,
+        },
+      });
+    },
+    [formSettings],
   );
 
   if (!sortedParams.length) {
@@ -59,6 +129,8 @@ function FormCreator({
       </SidebarContent>
     );
   }
+
+  const fieldSettings = formSettings.fields || {};
 
   const docsLink = (
     <ExternalLink
@@ -73,14 +145,45 @@ function FormCreator({
         <InfoText>
           {jt`Configure your parameters' types and properties here. The values for these parameters can come from user input, or from a dashboard filter. ${docsLink}`}
         </InfoText>
-        <ActionForm
-          parameters={sortedParams}
-          isEditable={isEditable}
-          onClose={_.noop}
-          onSubmit={_.noop}
-          formSettings={formSettings}
-          setFormSettings={setFormSettings}
-        />
+        <FormProvider
+          enableReinitialize
+          initialValues={BLANK_INITIAL_VALUES}
+          onSubmit={ON_SUBMIT_NOOP}
+        >
+          <Form role="form" data-testid="action-form-editor">
+            <DragDropContext onDragEnd={handleDragEnd}>
+              <Droppable droppableId="action-form-droppable">
+                {(provided: DroppableProvided) => (
+                  <div {...provided.droppableProps} ref={provided.innerRef}>
+                    {form.fields.map((field, index) => (
+                      <Draggable
+                        key={`draggable-${field.name}`}
+                        draggableId={field.name}
+                        isDragDisabled={!isEditable}
+                        index={index}
+                      >
+                        {(provided: DraggableProvided) => (
+                          <FormFieldEditorDragContainer
+                            ref={provided.innerRef}
+                            {...provided.draggableProps}
+                            {...provided.dragHandleProps}
+                          >
+                            <FormFieldEditor
+                              field={field}
+                              fieldSettings={fieldSettings[field.name]}
+                              isEditable={isEditable}
+                              onChange={handleChangeFieldSettings}
+                            />
+                          </FormFieldEditorDragContainer>
+                        )}
+                      </Draggable>
+                    ))}
+                  </div>
+                )}
+              </Droppable>
+            </DragDropContext>
+          </Form>
+        </FormProvider>
       </FormContainer>
     </SidebarContent>
   );
